backend/requirements.txt
Django==4.2.29
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-filter==25.1
